Move otel-ingestion unit tests out of src into tests/ folder
